BonchFeed v0.1
Stable version of Bonchfeed. App is using MVP architecture.
Model is using Rome lib for gathering rss feed from net and AsyncTask to perform the gathering in background.
Presenter contains an adapter for RecyclerView and tag handler for description fields in RSS with tags in it. On Item Click app sends user to the feed source.
View is a simple main activity with the Refresh button.